removed warnings from code,  most of which dealt with int and NSInteger

+++ With_v0/ChooseEventLocationViewController.m
@@ -224,7 +224,7 @@
 
 #pragma mark - TableView Delegates
 //NOT SURE WHERE THESE WARNINGS CAME FROM?
--(int)tableView:(UITableView *)tableView numberOfRowsInSection:(NSInteger)section
+-(NSInteger)tableView:(UITableView *)tableView numberOfRowsInSection:(NSInteger)section
 {
     if(self.isSearching == NO)
     {
